requirements.txt
Django==1.10.4
future==0.16.0
gunicorn==19.6.0
line-bot-sdk==1.0.2
PyMySQL==0.7.9
requests==2.12.3
wheel==0.24.0
whitenoise==3.2.2
wit
pythainlp
